Split useFlexHeaderSettings into focused hooks/utils, split tests to match
Extracts filter/header/page CRUD, sync toggle, stored errors, sync status,
and import-settings logic out of the monolithic settings.ts hook into their
own files (useFilterOperations, useHeaderOperations, usePageOperations,
useSyncToggle, useStoredErrors, useSyncStatus, importSettings, schemas),
each taking a small params object. settings.ts now just wires these
together plus owns core state/persistence.

Test files are reorganized to mirror: pageMerge.test.ts (pure merge/
tombstone logic), filterValidation.test.ts, usePageOperations.test.ts,
importSettings.test.ts (now a direct unit test of importSettingsFile),
and usePageHistory.test.ts (undo/redo + history toggle), replacing the
former catch-all settings.test.ts.
