+++ src/utils/pageStorage.ts
@@ -1,6 +1,6 @@
 import type browser from "webextension-polyfill";
 import { SETTINGS_V3_META_KEY, PAGE_KEY_PREFIX, PAGE_TOMBSTONES_KEY } from "../constants";
-import type { Page, SettingsV3Meta } from "./settings";
+import type { Page, SettingsV3Meta } from "./schemas";
 import { normalizePage } from "./headers";
 import type { PageTombstone } from "./pageMerge";
